Fix/juno.cha/20230303

See merge request pjt/bsp/sms/sms-wells-front-app!184

+++ src/project/wells/service/pages/adrwork/WwsnfMonthManagementCustomerMgtM.vue
@@ -84,6 +84,11 @@ const customCodes = {
  *  Event - 생성 버튼 클릭
  */
 async function onClickCreate() {
+  if (true) {
+    await alert('배치 개발 중입니다.');
+    return;
+  }
+
   if (!await confirm(t('MSG_ALT_IS_CRT_DATA'))) { return; }
   await dataService.post('/sms/wells/service/month-management', searchParams.value);
   // await notify(t('MSG_ALT_CREATED'));
@@ -93,6 +98,10 @@ async function onClickCreate() {
  *  Event - 삭제 버튼 클릭
  */
 async function onClickDelete() {
+  if (true) {
+    await alert('배치 개발 중입니다.');
+    return;
+  }
   if (!await confirm(t('MSG_ALT_DEL'))) { return; }
   await dataService.delete('/sms/wells/service/month-management', { data: searchParams.value });
   // await notify(t('MSG_ALT_DELETED'));
